@@ -46,11 +46,11 @@ module Moab
 
     # @attribute
     # @return [String] The digital object identifier (druid)
-    attribute :digital_object_id, String, :tag => 'objectId'
+    attribute :digital_object_id, String, tag: 'objectId'
 
     # @attribute
     # @return [Integer] The ordinal version number
-    attribute :version_id, Integer, :tag => 'versionId', :key => true, :on_save => proc { |n| n.to_s }
+    attribute :version_id, Integer, tag: 'versionId', key: true, on_save: proc { |n| n.to_s }
 
     # @return [String] The unique identifier concatenating digital object id with version id
     def composite_key
@@ -59,7 +59,7 @@ module Moab
 
     # @attribute
     # @return [String] The datetime at which the inventory was created
-    attribute :inventory_datetime, String, :tag => 'inventoryDatetime'
+    attribute :inventory_datetime, String, tag: 'inventoryDatetime'
 
     def inventory_datetime=(datetime)
       @inventory_datetime = Moab::UtcTime.input(datetime)
@@ -71,7 +71,7 @@ module Moab
 
     # @attribute
     # @return [Integer] The total number of data files in the inventory (dynamically calculated)
-    attribute :file_count, Integer, :tag => 'fileCount', :on_save => proc { |t| t.to_s }
+    attribute :file_count, Integer, tag: 'fileCount', on_save: proc { |t| t.to_s }
 
     def file_count
       groups.inject(0) { |sum, group| sum + group.file_count }
@@ -79,7 +79,7 @@ module Moab
 
      # @attribute
     # @return [Integer] The total size (in bytes) in all files of all files in the inventory (dynamically calculated)
-    attribute :byte_count, Integer, :tag => 'byteCount', :on_save => proc { |t| t.to_s }
+    attribute :byte_count, Integer, tag: 'byteCount', on_save: proc { |t| t.to_s }
 
     def byte_count
       groups.inject(0) { |sum, group| sum + group.byte_count }
@@ -87,7 +87,7 @@ module Moab
 
     # @attribute
     # @return [Integer] The total disk usage (in 1 kB blocks) of all data files (estimating du -k result) (dynamically calculated)
-    attribute :block_count, Integer, :tag => 'blockCount', :on_save => proc { |t| t.to_s }
+    attribute :block_count, Integer, tag: 'blockCount', on_save: proc { |t| t.to_s }
 
     def block_count
       groups.inject(0) { |sum, group| sum + group.block_count }
@@ -95,7 +95,7 @@ module Moab
 
     # @attribute
     # @return [Array<FileGroup>] The set of data groups comprising the version
-    has_many :groups, FileGroup, :tag => 'fileGroup'
+    has_many :groups, FileGroup, tag: 'fileGroup'
 
     # @return [Array<FileGroup] The set of data groups that contain files
     def non_empty_groups
@@ -194,7 +194,7 @@ module Moab
       signatures_from_bag = signatures_from_bagit_manifests(bag_pathname)
       bag_data_subdirs = bag_pathname.join('data').children
       bag_data_subdirs.each do |subdir|
-        groups << FileGroup.new(:group_id => subdir.basename.to_s).group_from_bagit_subdir(subdir, signatures_from_bag)
+        groups << FileGroup.new(group_id: subdir.basename.to_s).group_from_bagit_subdir(subdir, signatures_from_bag)
       end
       self
     end
