@@ -13,7 +13,7 @@ describe "Validate inventory against directory" do
     v1_inventory_pathname = @fixtures.join('derivatives/ingests/jq937jp0017/v0001/manifests/versionInventory.xml')
     v1_inventory = Moab::FileInventory.parse(v1_inventory_pathname.read)
     opts = {}
-    directory_inventory = Moab::FileInventory.new(:type => 'directory').inventory_from_directory(v1_data_directory)
+    directory_inventory = Moab::FileInventory.new(type: 'directory').inventory_from_directory(v1_data_directory)
     diff = Moab::FileInventoryDifference.new(opts)
     diff.compare(v1_inventory, directory_inventory)
     expect(diff.difference_count).to eq(0)
