@@ -13,7 +13,7 @@ describe "Build a version metadata datastream" do
     # given:
     # action:
     # outcome:
-    vm = Moab::VersionMetadata.new(:digital_object_id => 'druid:ab123cd4567')
+    vm = Moab::VersionMetadata.new(digital_object_id: 'druid:ab123cd4567')
 
     vme = Moab::VersionMetadataEntry.new
     vme.version_id = 1
@@ -21,32 +21,32 @@ describe "Build a version metadata datastream" do
     vme.description = 'Preserve my data please'
     vme.note = 'Initial object submission'
     vme.content_changes = Moab::FileGroupDifference.new(
-      :group_id => 'content',
-      :identical => 20,
-      :renamed => 2,
-      :modified => 0,
-      :added => 1,
-      :deleted => 0
+      group_id: 'content',
+      identical: 20,
+      renamed: 2,
+      modified: 0,
+      added: 1,
+      deleted: 0
     )
     vme.metadata_changes = Moab::FileGroupDifference.new(
-      :group_id => 'metadata',
-      :identical => 4,
-      :renamed => 0,
-      :modified => 3,
-      :added => 0,
-      :deleted => 0
+      group_id: 'metadata',
+      identical: 4,
+      renamed: 0,
+      modified: 3,
+      added: 0,
+      deleted: 0
     )
     vme.events << Moab::VersionMetadataEvent.new(
-      :type => "reopen",
-      :datetime => Time.now
+      type: "reopen",
+      datetime: Time.now
     )
     vme.events << Moab::VersionMetadataEvent.new(
-      :type => "submit",
-      :datetime => Time.now
+      type: "submit",
+      datetime: Time.now
     )
     vme.events << Moab::VersionMetadataEvent.new(
-      :type => "accession",
-      :datetime => Time.now
+      type: "accession",
+      datetime: Time.now
     )
 
     vm.versions << vme
