@@ -55,7 +55,7 @@ describe "Feature: Manifest Serialization" do
       </signatureCatalog>
     XML
     xmlObj2 = Nokogiri::XML(catalog_test)
-    opts = { :element_order => false, :normalize_whitespace => true }
+    opts = { element_order: false, normalize_whitespace: true }
     expect(EquivalentXml.equivalent?(xmlObj1, xmlObj2, opts)).to be true
   end
 end
