@@ -45,7 +45,7 @@ describe "Write contentMetadata datastream" do
       </contentMetadata>
     XML
     xmlObj2 = Nokogiri::XML(cm_test)
-    opts = { :element_order => false, :normalize_whitespace => true }
+    opts = { element_order: false, normalize_whitespace: true }
     expect(EquivalentXml.equivalent?(xmlObj1, xmlObj2, opts)).to be true
   end
 end
