@@ -18,7 +18,7 @@ describe "Export a digital object version from DOR" do
       inventory = Moab::FileInventory.read_xml_file(@manifests.join(vname), 'version')
       catalog = case version
                 when 1
-                  Moab::SignatureCatalog.new(:digital_object_id => inventory.digital_object_id)
+                  Moab::SignatureCatalog.new(digital_object_id: inventory.digital_object_id)
                 else
                   Moab::SignatureCatalog.read_xml_file(@manifests.join(@vname[version - 1]))
                 end
