@@ -84,7 +84,7 @@ describe Moab::Bagger do
           inventory = Moab::FileInventory.read_xml_file(@manifests.join(vname), 'version')
           catalog = case version
                     when 1
-                      Moab::SignatureCatalog.new(:digital_object_id => inventory.digital_object_id)
+                      Moab::SignatureCatalog.new(digital_object_id: inventory.digital_object_id)
                     else
                       Moab::SignatureCatalog.read_xml_file(@manifests.join(@vname[version - 1]))
                     end
