@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
 describe Stanford::StorageServices do
-  let(:eq_xml_opts) { { :element_order => false, :normalize_whitespace => true } }
+  let(:eq_xml_opts) { { element_order: false, normalize_whitespace: true } }
   let(:content_metadata) { IO.read(@data.join('v0002/metadata/contentMetadata.xml')) }
 
   specify '.cm_remediate' do
